requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
celery==5.3.0
redis==5.0.1
neo4j==5.13.0
qdrant-client==2.7.0
sentence-transformers==2.2.2
pydantic==2.5.0
python-multipart==0.0.6
pdfplumber==0.10.0
pillow==10.1.0
requests==2.31.0
numpy==1.24.3
transformers==4.35.0
torch==2.1.0
PyMuPDF==1.23.5
scikit-learn==1.3.0
